Implement a intermediate language for generation purposes
The google docs format, while much easier to parse than a HTML document,
is still not organized in an easy to use way for all the generation
targets we want to target. So, parse the JSON structure down to a
smaller AST-like intermediate structure, then generate from that with a
series of properties for each element type.

This enables testing as well; while not integrated yet it should be
easier to test the generation of documents from IL, and the parsing of
IL from gdocs independently.

Despite being a rewrite, we are largely close to being compatible
output-wise with our original sources, a few nits to correct later.

closes #12

Signed-off-by: Erik Hollensbe <github@hollensbe.org>